version which closely matches operational values when using original PET
tested using climate divisions, matching to monthly operational results
very closely with the original PET method used in place of the function
which uses the Thornthwaite method
