+++ pdinew.py
@@ -5,13 +5,15 @@ import numba
 import numpy as np
 import pandas as pd
 import profile
+import scipy
 import utils
 import warnings
 
 #-----------------------------------------------------------------------------------------------------------------------
 # list the objects that we'll make publicly visible from this module, as interpreted by 'import *'
 # a good explanation of this: https://stackoverflow.com/questions/44834/can-someone-explain-all-in-python
-__all__ = ['pdsi_from_climatology']
+__all__ = ['pdsi_from_climatology',
+           'potential_evapotranspiration']
 
 #-----------------------------------------------------------------------------------------------------------------------
 # global variables
@@ -1310,3 +1312,76 @@ def _case(PROB,
         PDSI = X3
  
     return PDSI
+
+#-----------------------------------------------------------------------------------------------------------------------
+# from pdinew.f
+def _pe(temperature,
+        month_index,
+        latitude,
+        data_start_year,
+        B,
+        H):
+    """
+    Computes potential evapotranspiration based on the method used in original PDSI code pdi.f
+    
+    :param temperature: temperature value in degrees Fahrenheit
+    :param month_index: index into monthly timeseries, with 0 corresponding to January of the initial year
+    :param latitude: latitude value in degrees north 
+    :param data_start_year: initial year of the data being computed
+    :param B: ?
+    :param H: ?
+    """
+    #TODO document this, where do these come from?
+    PHI = np.array([-0.3865982, -0.2316132, -0.0378180, 0.1715539, 0.3458803, 0.4308320, \
+                     0.3916645, 0.2452467, 0.0535511, -0.15583436, -0.3340551, -0.4310691])
+    
+    TLA = -1 * math.tan(math.radians(latitude))
+#     TLA = math.tan(math.radians(latitude))
+    
+    #-----------------------------------------------------------------------
+    #     1 - CALCULATE PE (POTENTIAL EVAPOTRANSPIRATION)   
+    #-----------------------------------------------------------------------
+    if temperature <= 32.0:
+        PE = 0.0
+    else:  
+        DUM = PHI[month_index % 12] * TLA 
+        
+        try:
+            DK = math.atan(math.sqrt(1.0 - (DUM * DUM)) / DUM)   
+        except ValueError:
+            logger.exception('Failed to complete', exc_info=True)
+            raise
+            
+        if DK < 0.0:
+            DK = 3.141593 + DK  
+        DK = (DK + 0.0157) / 1.57  
+        if temperature >= 80.0:
+            PE = (math.sin((temperature / 57.3) - 0.166) - 0.76) * DK
+        else:  
+            DUM = math.log(temperature - 32.0)
+            PE = math.exp(-3.863233 + (B * 1.715598) - (B * math.log(H)) + (B * DUM)) * DK 
+
+        #-----------------------------------------------------------------------
+        #     CONVERT DAILY TO MONTHLY  
+        #-----------------------------------------------------------------------
+        year = data_start_year + int((month_index + 1) / 12)
+        month = month_index % 12
+        PE = PE * calendar.monthrange(year, month + 1)[1]
+
+    return PE
+
+#-----------------------------------------------------------------------------------------------------------------------
+def potential_evapotranspiration(monthly_temps_celsius,
+                                 latitude,
+                                 data_start_year,
+                                 B,
+                                 H):
+
+    # assumes monthly_temps_celsius, B, and H have same dimensions, etc.
+    
+    pet = np.full(monthly_temps_celsius.shape, np.NaN)
+    monthly_temps_fahrenheit = scipy.constants.convert_temperature(monthly_temps_celsius, 'C', 'F')
+    for i in range(monthly_temps_celsius.size):
+        pet[i] = _pe(monthly_temps_fahrenheit[i], i, latitude, data_start_year, B, H)
+    return pet
+
